+++ core/src/main/java/com/cmeza/spring/excel/repository/resolvers/ExcelPropertyResolver.java
@@ -3,5 +3,9 @@ package com.cmeza.spring.excel.repository.resolvers;
 public interface ExcelPropertyResolver {
     String resolvePlaceholders(String key);
 
+    String resolvePlaceholders(String key, String message);
+
     String resolveRequiredPlaceholders(String key);
+
+    String resolveRequiredPlaceholders(String key, String message);
 }